requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
openai==1.3.7
beautifulsoup4==4.12.2
lxml==4.9.3
playwright==1.40.0
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.2

